requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.11
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.1
django-imagekit==5.0.0
django-js-asset==3.0.1
django-redis==5.4.0
django-taggit==5.0.1
django-taggit-labels==0.7.0
environs==14.1.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
magic-filter==1.0.12
Markdown==3.6
marshmallow==3.26.0
multidict==6.1.0
packaging==24.2
pilkit==3.0
pillow==11.1.0
pipdeptree==2.24.0
propcache==0.2.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.8.2
yarl==1.18.3
